feat(wmg): 3991 SVG download with MutationObserver
